# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OfferMaker API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/job-offers": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Get list of job offers",
                "operationId": "b52c2dff1cd7420ddb9c100bb00de415",
                "responses": {
                    "200": {
                        "description": "List of job offers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobOffer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Create a new job offer",
                "operationId": "383e512b0584bb406df0fc73c363cefc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "location",
                                    "contract_type",
                                    "salary"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Developer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "We are looking for a senior developer..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "contract_type": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "freelance"
                                        ],
                                        "example": "full-time"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 80000
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "closed"
                                        ],
                                        "example": "published"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job offer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobOffer"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/job-offers/{id}": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Get a specific job offer",
                "operationId": "ede56ac77d57187e58a7013898470e74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job offer ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job offer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobOffer"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Job offer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Update a job offer",
                "operationId": "65c7e25edcd77704e5f676ffa6121843",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job offer ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "location",
                                    "contract_type",
                                    "salary"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Developer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "We are looking for a senior developer..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "contract_type": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "freelance"
                                        ],
                                        "example": "full-time"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 80000
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "closed"
                                        ],
                                        "example": "published"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job offer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobOffer"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Job offer not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Delete a job offer",
                "operationId": "7ee0e1919ad3e6cf38ac7b2812907836",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job offer ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job offer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Job offer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "JobOffer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "recruiter_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "contract_type": {
                        "type": "string",
                        "enum": [
                            "full-time",
                            "part-time",
                            "freelance"
                        ]
                    },
                    "salary": {
                        "type": "number",
                        "format": "float"
                    },
                    "posted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "closed"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "recruiter": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "applications_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}